PoC in Fedora container
The PyPI installation gives random exceptions on ubuntu, but
if installed in a fedora container with dnf, it works as expected.

The projects list is still somewhat cumbersome, but if a toml is created
from an SBOM it works.

Graceful error handling isn't implemented, so the docker restart policy
is assumed.

Still, this works as a PoC, the release message is relayed to a teams
channel.